also demo remote chunk reader
